+++ pkg/tsdb/cloudwatch/models/types.go
@@ -1,0 +1,22 @@
+package models
+
+type cloudWatchLink struct {
+	View    string        `json:"view"`
+	Stacked bool          `json:"stacked"`
+	Title   string        `json:"title"`
+	Start   string        `json:"start"`
+	End     string        `json:"end"`
+	Region  string        `json:"region"`
+	Metrics []interface{} `json:"metrics"`
+}
+
+type metricExpression struct {
+	Expression string `json:"expression"`
+	Label      string `json:"label,omitempty"`
+}
+
+type metricStatMeta struct {
+	Stat   string `json:"stat"`
+	Period int    `json:"period"`
+	Label  string `json:"label,omitempty"`
+}
